Close #12: Provide finer load options
5 different load options are now available
v.low, low, mid, high and v.high.
The min and max values are unchanged [0, 2],
so future data should remain comparable.

+++ analyser.py
@@ -75,7 +75,7 @@ def parse_duration_string(dur_str):
 
 def parse_load_string(load_str):
     try:
-        return int(load_str)
+        return float(load_str)
     except ValueError:
         load_dict = {
             "low": 0,
